refactor: Move matrix/vector to new lib 'oxide_math'
