--- oxide_core/lib.rs
@@ -1,3 +1,0 @@
-pub mod matrix;
-pub mod test;
-pub mod vector;
