--- oxide_core/test.rs
@@ -1,3 +1,0 @@
-#[macro_use]
-pub mod utils;
-pub mod vector;
