--- oxide_core/test/vector.rs
@@ -1,3 +1,0 @@
-pub mod vector2;
-pub mod vector3;
-pub mod vector4;
