+++ oxide_math/lib.rs
@@ -1,0 +1,3 @@
+pub mod matrix;
+pub mod test;
+pub mod vector;
