+++ oxide_math/test.rs
@@ -1,0 +1,3 @@
+#[macro_use]
+pub mod utils;
+pub mod vector;
