+++ oxide_math/test/vector.rs
@@ -1,0 +1,3 @@
+pub mod vector2;
+pub mod vector3;
+pub mod vector4;
